move chart component files into chart-area folder
