+++ src/components/chart-area/BarChart.jsx
